herd-code/herd-service/src/main/java/org/finra/herd/service/RelationalTableRegistrationService.java
/*
* Copyright 2015 herd contributors
*
* Licensed under the Apache License, Version 2.0 (the "License");
* you may not use this file except in compliance with the License.
* You may obtain a copy of the License at
*
*     http://www.apache.org/licenses/LICENSE-2.0
*
* Unless required by applicable law or agreed to in writing, software
* distributed under the License is distributed on an "AS IS" BASIS,
* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
* See the License for the specific language governing permissions and
* limitations under the License.
*/
package org.finra.herd.service;

import org.finra.herd.model.api.xml.BusinessObjectData;
import org.finra.herd.model.api.xml.RelationalTableRegistrationCreateRequest;

/**
 * relational table registration service
 */
public interface RelationalTableRegistrationService
{
    /**
     * Create relational table registration in Herd
     * Includes create business object definition, business object format and business object data
     *
     * @param relationalTableRegistrationCreateRequest relational table registration create request
     *
     * @return business object data
     */
    public BusinessObjectData createRelationalTableRegistration(RelationalTableRegistrationCreateRequest relationalTableRegistrationCreateRequest);
}
